Basic HTML sections complete
Created main section to encompass the full controls, other sections are largely
self-explanatory by class name, except for app which describes the functional
area of the applet, i.e. everything that is not the static header.

Also added red 2px border to all items to help with wireframing the app.