Added submit list-item function
function takes the current value of the input and then creates a list item.
List item is filled with the necessary buttons and then added to the list
as a new first element (so that it is displayed first).
List item buttons currently have no real functionality, but have been
styled, details in commit to follow.